issue002-release0-print-command-line-parameters: Replace call to obsolete main function with call to a local skeleton main function
FEATURE(s):
issue002-release0-print-command-line-parameters:
  Parse the command line and print the environment name and the
  application name

FILE(s)
[modified]drl:
  Remove obsolete function calls

DESCRIPTION
Replace call to obsolete main function with call to a local skeleton
main function.

As part of the rewrite of the skeleton app, the first step is to
remove the call to obsolete configuration code and replace it with
a main function that is local to the drl script.

This will allow obsolete code to be deleted in a future commit.