issue002-release0-print-command-line-parameters: Move the test directory into the dralithus package
FEATURE(s):
issue002-release0-print-command-line-parameters:
  Parse the command line and print the environment name and the
  application name

FILE(s)
[renamed]test/.gitignore -> dralithus/test/.gitignore
[renamed]test/__init__.py -> dralithus/test/__init__.py
[renamed]test/configuration/.gitignore -> dralithus/test/configuration/.gitignore
[renamed]test/configuration/__init__.py -> dralithus/test/configuration/__init__.py
[renamed]test/configuration/process_command_line/.gitignore -> dralithus/test/configuration/process_command_line/.gitignore
[renamed]test/configuration/process_command_line/__init__.py -> dralithus/test/configuration/process_command_line/__init__.py
[renamed]test/configuration/process_command_line/args.py -> dralithus/test/configuration/process_command_line/args.py

DESCRIPTION
Move the test directory into the dralithus package

It seems like pylint within IntelliJ IDEA is not happy with the directory
structure. It is failing to find modules imported from within the test
directory. It clearly finds the modules from the command line and the
when run from the plugin console, so the problem is with the integration
of the plugin with the IDE.

It also seems a better practice to have the test in the same package as
the source code.