issue002-release0-print-command-line-parameters: Fix bug in merge_option_values that drops consecutive '-v' options
FEATURE(s):
issue002-release0-print-command-line-parameters:
  Parse the command line and print the environment name and the
  application name

FILE(s)
[modified]dralithus/configuration.py:
  Fix bug in merge_option_values that drops consecutive '-v' options

DESCRIPTION
Fix bug in merge_option_values that drops consecutive '-v' options.

Instead of adding the '-v' that precedes another '-v' option the
code that checks the next option for a value fails to add it to
the list of values when it catches a ValueError. Fix this by
adding the option to the list of values when the ValueError is
caught.