issue002-release0-print-command-line-parameters: Move make_args_list from test_help.py to __init__.py
FEATURE(s):
issue002-release0-print-command-line-parameters:
  Parse the command line and print the environment name and the
  application name

FILE(s)
[modified]dralithus/test/configuration/process_command_line/__init__.py:
  Add make_args_list function
[modified]dralithus/test/configuration/process_command_line/test_help.py:
  Remove make_args_list function

DESCRIPTION
make_args_list is a utility function that is used to create a list of
Arg object given lists of components of a command line. It effectively
create every possible combination of the components. This is useful
in creating test cases for the command line parser.

Since its utility is not limited to test_help.py, it is moved to
__init__.py so that it can be used by other test modules.