issue002-release0-print-command-line-parameters: Move the function that prints test cases to the module file
FEATURE(s):
issue002-release0-print-command-line-parameters:
  Parse the command line and print the environment name and the
  application name

FILE(s)
[modified]dralithus/test/configuration/process_command_line/__init__.py:
  Move the function that prints test cases to this module file.
  Modify it to take the test cases list as an argument.
[modified]dralithus/test/configuration/process_command_line/test_help.py:
  Remove the function that prints test cases from this module file.
  Use print_cases from the module file instead, and pass the test cases list
  as an argument.

DESCRIPTION
Move the function that prints test cases to the module file

This utility function is useful beyond just the test_help module, so it
makes sense to move it to the module file. This way, it can be used by
other modules as well.