issue002-release0-print-command-line-parameters: Define make_test_cases in command_line_parameters module
FEATURE(s):
issue002-release0-print-command-line-parameters:
  Parse the command line and print the environment name and the
  application name

FILE(s)
[modified]dralithus/test/configuration/process_command_line/__init__.py:
  Define make_test_cases

DESCRIPTION
Define make_test_cases in command_line_parameters module

This function is similar to the one in test_help, execpt that
it takes two additional arguments: expected and error which describe
the expected output and error output. It then iterates through the
args list to create one test case for each element in that args list
with the expected and error results.

This functon is different from a function of the same name in test_help,
because the latter function creates different error and expected objects
based on actual contents in each Arg object. This function does not.
So in a sense it is simpler.