issue002-release0-print-command-line-parameters: Modify type hints for make_test_cases to allow for None to be passed for expected and error parameters.
FEATURE(s):
issue002-release0-print-command-line-parameters:
  Parse the command line and print the environment name and the
  application name

FILE(s)
[modified]dralithus/test/configuration/process_command_line/__init__.py:
  Change type hints top allow the expected and error parameters to
  take None values.
DESCRIPTION
Modify type hints for make_test_cases to allow for None to be passed
for expected and error parameters.

This allows expected to be set to None when an error is expected to be
raised and error to be set to None when no error is expected to be
raised.