issue002-release0-print-command-line-parameters: Define a collection of test cases that are valid invocations of drl where the help flags are specified as a global option.
FEATURE(s):
issue002-release0-print-command-line-parameters:
  Parse the command line and print the environment name and the
  application name

FILE(s)
[modified]dralithus/test/configuration/process_command_line/test_help.py:
  Define help_valid_global_option_test_cases() that defines valid invocations
  of drl where the help flags are specified as a global option.

DESCRIPTION
Define a collection of test cases that are valid invocations of drl where
the help flags are specified as a global option.

This is way to partition the set of test cases along the following dimension:
1. valid/invalid invocations
2. global help option/command specific help option.

So in this classification there will be 4 categories of test cases:
1. valid invocations with global help option
2. invalid invocations with global help option
3. valid invocations with command specific help option
4. invalid invocations with command specific help option

The rationale for this classification is that each category of test cases
will have a different expected result.
For case 1, the expected result is an operation object with command set to help
and the about attribute set to None.
For case 2, the expected result is a CommandLineError exception.
For case 3, the expected result is an operation object with command set to help
and the about attribute set to the command name.
For case 4, the expected result is a CommandLineError exception.

Additionally command specific help test cases can be generated from the
help test cases for the command itself, by interleaving the help flags in
the command line invocation.