issue002-release0-print-command-line-parameters: Define class CommandLineError
FEATURE(s):
issue002-release0-print-command-line-parameters:
  Parse the command line and print the environment name and the
  application name

FILE(s)
[new]]dralithus/configuration2.py:
  Define class CommandLineError

DESCRIPTION
Define class CommandLineError

CommandLineError is a sub-class of DralithusError. It
represents an error that occurs when there is a problem with the
command line arguments passed to the application. The application
catches this exception during command line processing and converts
it into help command that gets executed.