issue002-release0-print-command-line-parameters: Move definition of GlobalOptions and CommandLineOptions to command_line_parser.py
FEATURE(s):
issue002-release0-print-command-line-parameters:
  Parse the command line and print the environment name and the
  application name

FILE(s)
[modified]dralithus/configuration2.py:
  Move the definition of the GlobalOptions and CommandLineOptions
  classes to command_line_parser.py
[new]dralithus/command_line_parser.py:
  Move the definition of the GlobalOptions and CommandLineOptions
  to this file.

DESCRIPTION
Move definition of GlobalOptions and CommandLineOptions to
command_line_parser.py.

This is in preparation for the next step, which is to create a
separate parser class for the command line. This separates the
storage of the command line options from the parsing of the command
line options and makes the code easier to understand and maintain.