issue002-release0-print-command-line-parameters: Move definition of class CommandLineError from command_line.py to errors.py
FEATURE(s):
issue002-release0-print-command-line-parameters:
  Parse the command line and print the environment name and the
  application name

FILE(s)
[modified]dralithus/command_line.py:
  Move definition of class CommandLineError to errors.py
[modified]dralithus/errors.py:
  Move definition of class CommandLineError from command_line.py
  to this file

DESCRIPTION
Move definition of class CommandLineError from command_line.py
to errors.py.

This is done because the class CommandLine in file command_line.py (to
be defined in a future commit) as well as class CommandLineParser in
file command_line_parser.py (also to be defined in a future commit,)
both use the class CommandLineError.  However, the class CommandLine
will use class CommandLineParser. This will result in a circular
dependency.  To avoid this, the class CommandLineError is moved to
file errors.py.  errors.py is also a more appropriate place for
exception classes.