issue002-release0-print-command-line-parameters:
FEATURE(s):
issue002-release0-print-command-line-parameters:

FILE(s)
[modified]dralithus/command.py:
  Move the make function out of class Command and make it a module
  level function
[modified]dralithus/test/test_command.py:
  Call command.make() rather than Command.make()

[modified]dralithus/test/test_command.py

DESCRIPTION
Move the make function out of class Command and make it a module
level function.

This is done because the make requires different parameters in
Command.make and the subtypes of Command. As a classmethod or
static method in Command the signature would require a list
of strings (the command line args.) However, in subtype of
Command, the respective make functions require the program
name, command string, global and command options.

This difference in signature will result in pylint warnings
about this. To avoid those warnings, we simple move the make
function to the module level.