issue002-release0-print-command-line-parameters: Modify CaseExecutor2.execute() to treat [None] as an expected value of None.
FEATURE(s):
issue002-release0-print-command-line-parameters:
  Parse the command line and print the environment name and the
  application name

FILE(s)
modified:   dralithus/test/__init__.py
Modify CaseExecutor2.execute() to treat [None] as an expected value
of None.

DESCRIPTION
This fixes design bug in the CaseExecutor framework when the expected
return value of the function under test is None. In that case the code
had no way to distinguish between a function that returned None and a
function that dio not return anything, but instead raised an
exception. This change allows the CaseExecutor2 to handle the case
where the expected return value is None.