issue002-release0-print-command-line-parameters: dralithus/command_line/environment_option.py: do not check for existence of an environment when parsing the command line
FEATURE(s):
issue002-release0-print-command-line-parameters:
  Parse the command line and print the environment name and the
  application name

FILE(s)
[modified]dralithus/command_line/environment_option.py:
  do not check for existence of an environment when parsing the command line
[modified]dralithus/test/command_line/test_environment_option.py:
  modify tests to check that options with names of environments that
  don't exist are still considered valid environment options.
[modified]dralithus/test/command_line/test_option.py:
  modify tests to check that options with names of environments that
  don't exist are still considered valid options.
[modified]dralithus/test/command_line/test_options.py:
  modify tests to check that options with names of environments that
  don't exist are still considered valid options.

DESCRIPTION
dralithus/command_line/environment_option.py: do not check for
existence of an environment when parsing the command line. Also
modify tests to check that options with names of environments that
don't exist are still considered valid options. Earlier these
tests would check that that parsing failed if the environment
existed.

The purpose of this change is to restrict the check for the existence
of an environment to a single function in the code base (DRY principle.)
This should be within the code that creats the environment further
up the stack.